new migrations written. more comprehensive seeds needed

+++ server/db/migrations/20240326020716_item.js
@@ -8,7 +8,7 @@ export function up(knex) {
     table.string('name')
     table.text('desc')
     table.date('date_added')
-    table.numeric('weight', 7, 2)
+    table.integer('weight', 7, 2)
     table.string('location')
     table.integer('owner_id').nullable()
     table.string('image_src')
